refactor: move domain classes into separate file
