+++ src/app/chatbot/chatbot.service.ts
@@ -2,28 +2,7 @@ import { Injectable, inject } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { firstValueFrom, Observable, of } from 'rxjs';
 import { map, catchError } from 'rxjs/operators';
-
-export interface Chat {
-  id: string;
-  title: string;
-  messages: ChatMessage[];
-  createdAt: Date;
-  lastUpdated: Date;
-  deleted: boolean;
-}
-
-export class ChatMessage {
-  role: 'user' | 'assistant' | 'system';
-  content: string;
-  nodeId?: string;
-  timestamp: Date = new Date();
-
-  constructor(role: 'user' | 'assistant' | 'system', content: string, nodeId?: string) {
-    this.role = role;
-    this.content = content;
-    this.nodeId = nodeId;
-  }
-}
+import { Chat, ChatMessage } from './chat';
 
 @Injectable({ providedIn: 'root' })
 export class ChatbotService {
